requirements.txt
# Genesis Bots requirements file
annoy==1.17.3
Brotli==1.0.9 # required for annoy to work on virtualized Windows machines
APscheduler
bs4
db-dtypes
dill
protobuf==4.21.11
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
werkzeug>=3.0.6
flask==3.0.3
google-cloud-bigquery
mistralai
openai>=1.66.0
pandas==2.1.4 # needed < 2.2.2 for databricks connector to work
pyarrow==17.0.0
pytest
flask-cors==5.0.1
pyyaml==6.0.1
simplejson==3.19.2
slack_bolt
streamlit==1.35.0
streamlit_autorefresh
urllib3==1.26.19
ngrok
snowflake_connector_python==3.12.3
reka-api
google-generativeai
pdb_attach
sseclient
botbuilder-integration-aiohttp>=4.15.0
jira==3.8.0
gitpython==3.1.43
anthropic==0.40.0
boto3>=1.26.0
markdown
cython
tenacity
dagster==1.9.5
dagster-graphql
dagster-spark
dagster-dbt
wrapt
sqlalchemy>=1.4.0
psycopg2-binary==2.9.9
pymysql
snowflake-sqlalchemy
spider-client
openpyxl
sqlalchemy-redshift
sqlalchemy-bigquery
PyGithub
sqlalchemy-databricks
dbt-core
dbt-snowflake
pymupdf
jsonpickle
jmespath
llama-index
pypdf
sentence-transformers
xlrd
fastapi==0.115.11
chromadb
llama-index[chromadb]
llama-index-vector-stores-chroma
